+++ src/v1/imports/carts/buddi.ts
@@ -1,4 +1,4 @@
-import { Transactions } from "../../shared/types";
+import { Transactions } from "../../../shared/types";
 
 const buddiTracker = ({
   appId,
