+++ src/v1/imports/carts/dutchie-subdomain.ts
@@ -1,4 +1,4 @@
-import { Transactions } from "../../shared/types";
+import { Transactions } from "../../../shared/types";
 
 const dutchieSubdomainTracker = ({
   appId,
