+++ src/v1/imports/carts/tymber.ts
@@ -1,4 +1,4 @@
-import { Transactions } from "../../shared/types";
+import { Transactions } from "../../../shared/types";
 
 const tymberTracker = ({
   appId,
