+++ src/v1/imports/impressions/liquidm.ts
@@ -43,9 +43,9 @@ export const liquidm = () => {
   };
 
   mjcx.push(cCx),
-  mjcx.push(cCx2),
-  mjcx.push(cCx3),
-  window.tracker("trackSelfDescribingEvent", unstruct, mjcx);
+    mjcx.push(cCx2),
+    mjcx.push(cCx3),
+    window.tracker("trackSelfDescribingEvent", unstruct, mjcx);
 }
 
 export default liquidm;
